requirements.txt
numba==0.49.0
numpy==1.18.3
Pillow==7.1.2
